Uploading Assignments.
With this commit, I am uploading assignments. However, since there is no matching of assignments to week number, I am not able to pinpoint which assignment belongs to which week.
Since I feel that these assignments belong to week 5, I am uploading the same in week5 folder.

I also noticed that there is a blank file in each of the week folder. This blank has has no content and it does not even have a file extension. Perhaps these are system created files. The need not exist in the folders and have no purpose. Hence I am deleting those files with this commit.